+++ CHANGELOG.md
@@ -26,6 +26,7 @@ versioned in accordance with [Semantic Versioning](https://semver.org/spec/v2.0.
 - BaseConfiguration now provides a `yaml` method for easily to YAML.
 - BaseConfiguration now provides a `json_encoders` static method for easily accessing the default encoders.
 - The Vegeta connector now validates the `target` and `targets` settings.
+- The check command now supports a verbose and non-verbose modes for outputting multiple checks from a connector.
 
 ### Removed
 - The `durationpy` package has been removed in favor of a local implementation.
@@ -34,6 +35,7 @@ versioned in accordance with [Semantic Versioning](https://semver.org/spec/v2.0.
 - Updated Docker and Docker Compose configurations to use `/servo/opsani.token` as the default path for API tokens mounted as a file.
 - Docker images pushed to Docker Hub are now built with `SERVO_ENV=production` to exclude development packages.
 - The reporting interval is now configurable `VegetaConfiguration`.
+- The check event now returns a list of checks instead of one result.
 
 ## [0.3.1] - 2020-07-16